refactor: refactor bot module
