+++ src/environment/discord-config.service.ts
